Move plugin functional helper code to main func helper file
